diff kernel

+++ scratch/kernels/diff/diff.c
@@ -1,5 +1,5 @@
 #include <stdint.h>
-void func(short*  s, int*  t, int n) {
+void func(int*  s, int*  t, int n) {
   for (int i = 0 ; i < n; i +=2) {
     t[i/2] = s[i] - s[i + 1];
   }
